+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+- Add `-v, --version` command for cli to print current version
+
 ## v0.7.0 - 07.01.25
 ### Fixed
 - Docstrings and types of CpxE4Iol `write_channel()`